Refactoring DietaryRestrictionDataAccessObject
